add test cycles api
